% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_norm_counts.R
\name{calc_norm_counts}
\alias{calc_norm_counts}
\title{Filter, normalize, and/or convert counts object}
\usage{
\code{
calc_norm_counts(
  counts, design, libID_col="lib.id",
  min_count=NULL, min_cpm=NULL, min_libs_perc=0.15,
  normalize=TRUE, norm_method="TMM",
  log2_transform=FALSE, transpose=FALSE, return_DGEcounts=FALSE,
  ...)}
}
\arguments{
\item{counts}{a matrix or data frame of gene expression counts. Should have sample in columns and genes in rows.}

\item{design}{a data frame of sample information. At minimum, must contain a column corresponding to sample identifiers matching column names of \code{counts}. Passed to \code{design_filter_counts}.}

\item{libID_col}{numeric index or character name of column in \code{design} containing sample identifers matching column names of \code{counts}. Passed to \code{design_filter_counts}.}

\item{min_count}{numeric, the minimum count for a library to be considered passing threshold for a given gene.}

\item{min_cpm}{numeric, the minimum counts per million for a library to be considered passing threshold for a given gene.}

\item{min_libs_perc}{numeric, the minimum percentage of libraries that must meet \code{min_count} or \code{min_cpm} threshold for a gene to be retained.}

\item{normalize}{logical, whether to normalize counts using \code{edgeR::calcNormFactors}.}

\item{norm_method}{character, the method by which to normalize counts; passed to \code{edgeR::calcNormFactors}. Used only if \code{normalize} is TRUE. Defaults to "TMM".}

\item{log2_transform}{logical, whether to log2 transform the counts. Defaults to \code{FALSE}.}

\item{transpose}{logical, whether to transpose the matrix or data frame of counts.}

\item{return_DGEcounts}{logical, whether to return return counts as a \code{DGEList} object. If FALSE, counts are returned as a data frame. Defaults to FALSE.}

\item{...}{(optional) parameters passed to \code{DGEList}.}
}
\value{
a data frame or \code{DGEList} object containing the processed counts.
}
\description{
This function performs a number of operations on a counts object. It filters the counts to include
only samples matching a sample annotation object. It optionally filters out low-count genes. It
optionally normalizes, log2-transforms, and/or transposes counts. Finally, it can return either
a DGEList object or a data frame.
}
\details{
This function utilizes \code{design_filter_counts} and (optionally) \code{min_filter_counts}
 to filter the counts object. It then (optionally) normalizes the counts, using
 \code{edgeR::calcNormFactors}. It then (optionally) log2-transforms the counts and/or transposes
 the counts object. Finally, it returns the counts either as a data frame, or as a \code{DGEList}
 object.
 For WGCNA analyses, should use \code{log2_transform=TRUE, transpose=TRUE, return_DGECounts=FALSE}.
 For limma analyses, should use \code{log2_transform=FALSE, transpose=FALSE, return_DGECounts=TRUE}.
}
