% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_norm_counts_scRNAseq.R
\name{calc_norm_counts_scRNAseq}
\alias{calc_norm_counts_scRNAseq}
\title{Filter, normalize, and/or convert counts object for single-cell RNA-seq data}
\usage{
\code{
calc_norm_counts_scRNAseq(
  counts,
  min_count=NULL, min_cpm=NULL, min_libs_perc=0.15,
  normalize=TRUE, norm_method="TMM",
  log2_transform=FALSE, transpose=FALSE, return_DGEcounts=FALSE,
  ...)}
}
\arguments{
\item{counts}{a matrix or data frame of gene expression counts. Should have samples in columns and genes in rows.}

\item{min_count}{numeric, the minimum count for a library to be considered passing threshold for a given gene. Setting to NULL avoids filtering by minimum counts.}

\item{min_cpm}{numeric, the minimum counts per million for a library to be considered passing threshold for a given gene. Setting to NULL avoids filtering by minimum counts.}

\item{min_libs_perc}{numeric, the minimum percentage of libraries that must meet \code{min_count} or \code{min_cpm} threshold for a gene to be retained.}

\item{normalize}{logical, whether to normalize counts.}

\item{norm_method}{character, the method by which to normalize counts. Used only if \code{normalize} is TRUE. Options are "TMM", which uses \code{edgeR::calcNormFactors}; "deconvolution", which uses cell deconvolution as implemented in the package \code{scran}; or "lib_size", simple library size normalization using cpm. Unique partial matches are accepted.}

\item{log2_transform}{logical, whether to log2 transform the counts. Defaults to \code{FALSE}.}

\item{transpose}{logical, whether to transpose the matrix or data frame of counts.}

\item{...}{(optional) parameters passed to normalization functions.}
}
\value{
a data frame containing the processed counts.
}
\description{
This function performs a number of operations on a counts object. It optionally filters out low-count genes. It
optionally normalizes, log2-transforms, and/or transposes counts.
}
\details{
This function (optionally) utilizes \code{min_filter_counts}
 to filter the counts object. It then (optionally) normalizes the counts, using
 one of several methods applicable to single-cell RNA-seq data. It then (optionally)
 log2-transforms the counts and/or transposes the counts object. Finally, it returns
 the counts either as a data frame, or as a \code{DGEList} object.
}
