% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/design_filter_counts.R
\name{design_filter_counts}
\alias{design_filter_counts}
\title{Remove samples without annotation data from counts object}
\usage{
\code{design_filter_counts(counts, design, libID_col)}
}
\arguments{
\item{counts}{a matrix or data frame of gene expression counts. Must have samples in columns and genes in rows, with colnames as sample identifiers.}

\item{design}{a data frame of sample information. At minimum, must contain a column corresponding to sample identifiers matching column names of \code{counts}.}

\item{libID_col}{numeric index or character name of column in \code{design} containing sample identifers matching column names of \code{counts}.}
}
\value{
a matrix or data frame with the same number of rows as \code{counts}
}
\description{
This function filters a counts object, removing samples that are not found in a corresponding
sample annotation object.
}
\details{
Samples are removed from \code{counts} if their column name does not match an element in column \code{libID_col} of \code{design}.
}

