% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_counts.R
\name{extract_counts}
\alias{extract_counts}
\title{Extract counts matrix from different types of expression objects}
\usage{
\code{
extract_counts(counts, return_class=NULL)}
}
\arguments{
\item{counts}{an object from which counts can be extracted. May be of class matrix, data.frame, EList, DGElist, ExpressionSet, eSet.}

\item{return_class}{optional, the class for the object to be returned. Defaults to NULL, which causes the counts portion of the object to be returned as-is.}
}
\value{
A matrix or data frame.
}
\description{
This is a helper function to extract the counts matrix from different gene expression object classes. It is used
by many other functions to enable counts being input as matrices, data.frames, EList or DGEList objects. It
determines the class of the object and returns the counts as a data.frame (default) or matrix.
}
