% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minFilterCounts.R
\name{minFilterCounts}
\alias{minFilterCounts}
\title{Filter low-count genes}
\usage{
\code{minFilterCounts(counts, min_count=1, min_libs_perc=0.15)}
}
\arguments{
\item{counts}{a matrix or data frame of gene expression counts; should have samples in columns and genes in rows.}

\item{min_count}{numeric value, the minimum count value}

\item{min_libs_perc}{numeric value, the minimum percentage of libraries}
}
\value{
a matrix or data frame with the same number of columns as \code{counts}
}
\description{
This function filters a counts object, removing genes that do not meet a minimum count value in
a minimum percentage of libraries
}
\details{
Genes are removed from \code{counts} if they do not have at least \code{min_count} counts in at least \code{min_libs_perc} % of all libraries.
 Note that adding or removing libraries to/from the counts object may change the genes that meet these thresholds.
}

