% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/var_filter_counts.R
\name{var_filter_counts}
\alias{var_filter_counts}
\title{Filter low-variance genes}
\usage{
\code{
var_filter_counts(
  counts, counts, n_keep=8000, genes="rows",
  log2_transform=FALSE)}
}
\arguments{
\item{counts}{a matrix or data frame of gene expression counts.}

\item{n_keep}{integer, the maximum number of genes to retain}

\item{genes}{character string, the dimension of \code{counts} containing the genes. Accepted values are "rows", "columns", and partial matches. Defaults to "rows"}

\item{log2_transform}{boolean, whether to log2 transform the counts prior to variance filtering. If the counts have not already been transformed, should be set to \code{TRUE}. Defaults to \code{FALSE}.}
}
\value{
a matrix or data frame with the same number of samples as \code{counts}, but possibly with fewer genes.
}
\description{
This function filters a counts object, keeping the \code{n_keep} genes with the highest variance.
}
\details{
Variance is calculated for all genes. Genes are then ranked by this variance, and the \code{n_keep} genes with highest variance are retained.
}

